cc-v3/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
python-dotenv==1.0.1
openpyxl==3.1.5
pydantic==2.9.2
streamlit==1.38.0
pandas==2.2.3
requests==2.32.3
